WIP: Initial work around making a checksum-reader
The checksum reader supports either seeking or buffering in memory and
also caches checksums.

The intent is that if a certain file or piece of data wants to be used
with multiple services, the checksum can be calculated only once, thus
saving cpu.

This is a WORK IN PROGRESS and does not work in any sense of the word.